Splitted Reply and Request from tier-node
